Auto-generate: Move promisify to utils
Fixes #999
