+++ lib/config/generator.js
@@ -158,27 +158,8 @@ Generator.prototype._getUserViolationChoices = function(errorPrompts) {
     return this._showPrompt(errorPrompts);
 };
 
-function promisify(fn) {
-    return function() {
-        var deferred = Vow.defer();
-        var args = [].slice.call(arguments);
-
-        args.push(function(err, result) {
-            if (err) {
-                deferred.reject(err);
-            } else {
-                deferred.resolve(result);
-            }
-        });
-
-        fn.apply(null, args);
-
-        return deferred.promise();
-    };
-}
-
 /** @private */
-Generator.prototype._showPrompt = promisify(prompt.get.bind(prompt));
+Generator.prototype._showPrompt = utils.promisify(prompt.get.bind(prompt));
 
 /**
  * @private
