+++ lib/utils.js
@@ -1,4 +1,5 @@
 var path = require('path');
+var Vow = require('vow');
 
 // 7.5.2 Keywords
 var ES3_KEYWORDS = {
@@ -213,6 +214,33 @@ exports.normalizePath = function(filepath, basePath) {
     }
 
     return filepath;
+};
+
+/**
+ * Wraps a function such that you can interact with a promise and not a
+ * node-style callback.
+ *
+ * @param  {Function} fn - function that expects a node-style callback
+ * @return {Function} When invoked with arguments, returns a promise resolved/rejected
+ *                    based on the results of the wrapped node-style callback
+ */
+exports.promisify = function(fn) {
+    return function() {
+        var deferred = Vow.defer();
+        var args = [].slice.call(arguments);
+
+        args.push(function(err, result) {
+            if (err) {
+                deferred.reject(err);
+            } else {
+                deferred.resolve(result);
+            }
+        });
+
+        fn.apply(null, args);
+
+        return deferred.promise();
+    };
 };
 
 /**
